+++ src/gnuwin32/extra.c
@@ -1332,7 +1332,7 @@ int winAccess(const char *path, int mode)
     if(GA_isNT) {
 	/* Now look for file security info, which is NT only */
 	SECURITY_DESCRIPTOR *sdPtr = NULL;
-	unsigned long size = 0;
+	DWORD size = 0;
 	GENERIC_MAPPING genMap;
 	HANDLE hToken = NULL;
 	DWORD desiredAccess = 0;
